Move get_all_metrics method to BaseProm.
