+++ nops_k8s_agent/nops_k8s_agent/container_cost/base_prom.py
@@ -1,4 +1,5 @@
 import sys
+from collections import defaultdict
 from datetime import datetime
 from typing import Any
 
@@ -31,3 +32,18 @@ class BaseProm:
         except Exception as e:
             logger.error(f"Error in get_metrics: {e}")
             return None
+
+    def build_query(metric_name, step):
+        pass
+
+    def get_all_metrics(self, start_time: datetime, end_time: datetime, step: str) -> dict:
+        # This function to get all metrics from prometheus
+        metrics = defaultdict(list)
+        for metric_name in self.list_of_metrics.keys():
+            query = self.build_query(metric_name, step)
+            response = self.get_metrics(
+                query=query, start_time=start_time, end_time=end_time, metric_name=metric_name, step=step
+            )
+            if response:
+                metrics[metric_name] = response
+        return metrics
